backend/requirements.txt
flask
redis
psycopg2-binary
flask-cors
flask-swagger-ui